+++ src/acconeer/exptool/a121/algo/presence/_detector_plugin.py
@@ -18,6 +18,7 @@ import pyqtgraph as pg
 
 import acconeer.exptool as et
 from acconeer.exptool import a121
+from acconeer.exptool.a121._h5_utils import _create_h5_string_dataset
 from acconeer.exptool.a121.algo._base import AlgoConfigBase
 from acconeer.exptool.a121.algo._plugins import (
     DetectorBackendPluginBase,
@@ -66,25 +67,11 @@ class SharedState:
     replaying: bool = attrs.field(default=False)
 
 
-@attrs.frozen(kw_only=True)
-class Save(AlgoConfigBase):
-    config: DetectorConfig = attrs.field()
-    plot_config: PlotConfig = attrs.field()
-
-    def to_dict(self) -> dict[str, Any]:
-        return {"config": self.config.to_dict(), "plot_config": self.plot_config.to_dict()}
-
-    @classmethod
-    def from_dict(cls: type[Save], obj: dict[str, Any]) -> Save:
-        return Save(
-            config=DetectorConfig.from_dict(obj["config"]),
-            plot_config=PlotConfig.from_dict(obj["plot_config"]),
-        )
-
-
 class BackendPlugin(DetectorBackendPluginBase[SharedState]):
-    def __init__(self, callback: Callable[[Message], None], key: str) -> None:
-        super().__init__(callback=callback, key=key)
+    def __init__(
+        self, callback: Callable[[Message], None], generation: PluginGeneration, key: str
+    ) -> None:
+        super().__init__(callback=callback, generation=generation, key=key)
 
         self._started: bool = False
         self._live_client: Optional[a121.Client] = None
@@ -96,17 +83,15 @@ class BackendPlugin(DetectorBackendPluginBase[SharedState]):
         self.restore_defaults()
 
     @is_task
-    def _from_cache(self, *, data: dict) -> None:
-        obj = Save.from_dict(data)
-        self.shared_state.config = obj.config
-        self.shared_state.plot_config = obj.plot_config
-        self.broadcast(sync=True)
+    def load_from_cache(self) -> None:
+        try:
+            with self.h5_cache_file() as f:
+                self.shared_state.config = DetectorConfig.from_json(f["config"][()])
+                self.shared_state.plot_config = PlotConfig.from_json(f["plot_config"][()])
+        except FileNotFoundError:
+            pass
 
-    def _to_cache(self) -> dict:
-        return Save(
-            config=self.shared_state.config,
-            plot_config=self.shared_state.plot_config,
-        ).to_dict()
+        self.broadcast(sync=True)
 
     def broadcast(self, sync: bool = False) -> None:
         super().broadcast()
@@ -155,16 +140,15 @@ class BackendPlugin(DetectorBackendPluginBase[SharedState]):
         self.broadcast()
 
     def teardown(self) -> None:
-        self.callback(
-            GeneralMessage(
-                name="serialized",
-                kwargs={
-                    "generation": PluginGeneration.A121,
-                    "key": self.key,
-                    "data": self._to_cache(),
-                },
-            )
-        )
+        try:
+            with self.h5_cache_file(write=True) as f:
+                _create_h5_string_dataset(f, "config", self.shared_state.config.to_json())
+                _create_h5_string_dataset(
+                    f, "plot_config", self.shared_state.plot_config.to_json()
+                )
+        except Exception:
+            log.warning("Detector could not write to cache")
+
         self.detach_client()
 
     @is_task
@@ -849,7 +833,7 @@ class PluginSpec(PluginSpecBase):
     def create_backend_plugin(
         self, callback: Callable[[Message], None], key: str
     ) -> BackendPlugin:
-        return BackendPlugin(callback=callback, key=key)
+        return BackendPlugin(callback=callback, generation=self.generation, key=key)
 
     def create_view_plugin(self, app_model: AppModel, view_widget: QWidget) -> ViewPlugin:
         return ViewPlugin(app_model=app_model, view_widget=view_widget)